Remove OpenStack 1.0 support, collapsing into RS.
There's consensus that since Rackspace is the only public OpenStack
1.0-ish implementation, the Rackspace driver should be self-contained
for the time being. Things should look different for 1.1 going forward
due to wider adoption.
