+++ test/compute/fixtures/rackspace/v1_slug_flavors_detail.xml
