+++ test/compute/fixtures/rackspace/v1_slug_images_post.xml
