+++ test/compute/fixtures/rackspace/v1_slug_servers_detail_empty.xml
