+++ test/compute/fixtures/rackspace/v1_slug_shared_ip_groups.xml
